Fixed some warnings from Codacy

+++ modmult.c
@@ -1147,7 +1147,7 @@ void modmult(const limb *factor1, const limb *factor2, limb *product)
     {
       int32_t MontDig;
       int32_t Nbr = (factor1 + i)->x;
-      int64_t Pr = (int64_t)Nbr * (int64_t)factor2->x + (int64_t)Prod[0].x;
+      int64_t Pr = ((int64_t)Nbr * (int64_t)factor2->x) + (int64_t)Prod[0].x;
       MontDig = ((int32_t)Pr * MontgomeryMultN[0].x) & MAX_VALUE_LIMB;
       Pr = (((int64_t)MontDig * (int64_t)TestNbr[0].x + Pr) >> BITS_PER_GROUP) +
         ((int64_t)MontDig * (int64_t)TestNbr[1].x) + ((int64_t)Nbr * (int64_t)(factor2 + 1)->x) + (int64_t)Prod[1].x;
@@ -1177,7 +1177,8 @@ void modmult(const limb *factor1, const limb *factor2, limb *product)
       dAccum = floor((dAccum*dInvLimbRange) + 0.5);
       low = ((int)dAccum + (MontDig * TestNbr[1].x) +
                    (Nbr * (factor2 + 1)->x) + Prod[1].x) & MAX_VALUE_LIMB;
-      dAccum += dMontDig * TestNbr[1].x + dNbr * (factor2 + 1)->x + (unsigned int)Prod[1].x;
+      dAccum += (dMontDig * (double)TestNbr[1].x) + (dNbr * (double)(factor2 + 1)->x) +
+        (double)Prod[1].x;
       Prod[0].x = low;
       for (j = 2; j < NumberLength; j++)
       {
@@ -1191,10 +1192,11 @@ void modmult(const limb *factor1, const limb *factor2, limb *product)
         {
           dAccum = ((dAccum - (double)FOURTH_INT_RANGE)*dInvLimbRange);
         }
-        low = (int)(dAccum - floor(dAccum * dInvLimbRange) * dLimbRange);
-        dAccum += dMontDig * TestNbr[j].x + dNbr * (factor2 + j)->x + (unsigned int)Prod[j].x;
-        low = (low + MontDig * TestNbr[j].x +
-               Nbr * (factor2 + j)->x + Prod[j].x) & MAX_VALUE_LIMB;
+        low = (int)(dAccum - (floor(dAccum * dInvLimbRange) * dLimbRange));
+        dAccum += (dMontDig * (double)TestNbr[j].x) + (dNbr * (double)(factor2 + j)->x) +
+          (double)Prod[j].x;
+        low = (low + (MontDig * TestNbr[j].x) +
+               (Nbr * (factor2 + j)->x) + Prod[j].x) & MAX_VALUE_LIMB;
         Prod[j - 1].x = low;
       }
       if (low < HALF_INT_RANGE)
